refactor(media-query): use index file convention
