+++ src/components/Footer/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { mediaQueryWrapper, DesktopBreakSize } from '../MediaQuery/MediaQueryWrapper';
+import { mediaQueryWrapper, DesktopBreakSize } from '../MediaQueryWrapper';
 
 export const Footer: React.SFC = () => {
   return (
