+++ src/components/GenericPage/GenericLayout.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { PageWidthEnforcer } from './PageWidthEnforcer';
 import homeLogo from '../../assets/images/home-page.jpg';
 import { SubHeader } from '../MainContent/Components';
-import { MobileBreakSize } from '../MediaQuery/MediaQueryWrapper';
+import { MobileBreakSize } from '../MediaQueryWrapper';
 
 export const genericLayout = <T extends object>(
   title: React.ReactNode,
