+++ src/components/TitleBar/mobileTitleBar.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { DownloadBtn, DownloadBtnProps } from './DownloadBtn';
 import { Styles, Colors } from '../../constants/style';
 import { IDomain } from '../../models/claim';
-import { MobileBreakSize, mediaQueryWrapper } from '../MediaQuery/MediaQueryWrapper';
+import { MobileBreakSize, mediaQueryWrapper } from '../MediaQueryWrapper';
 
 /**
  * interface for TitleBar
